+++ spec/concepts/bloom_remit/txns/operations/update_spec.rb
@@ -14,15 +14,21 @@ module BloomRemit
             secret: txn.secret,
             status: "paid",
             status_description: "My status description",
+            vendor_external_id: "vextid",
             tracking_number: "tracking-number",
+            fake_id: "fake_id",
           )
           txn = op.model
           expect(txn.status).to eq "paid"
           expect(txn.status_description).to eq "My status description"
           expect(OnUpdateJob).to have_enqueued_job(txn.id, {
+            id: txn.id,
+            secret: txn.secret,
             status: "paid",
             status_description: "My status description",
+            vendor_external_id: "vextid",
             tracking_number: "tracking-number",
+            fake_id: "fake_id",
           })
         end